Moved smoke tests and updated the npm scripts
