+++ cypress/integration/smokeTests/prodOnly/basic-enduser-prod-only.ts
