+++ cypress/integration/smokeTests/sharedTests/monitor.ts
