+++ cypress/integration/smokeTests/sharedTests/waf.ts
