+++ package.json
@@ -21,15 +21,22 @@
     "lint": "npx ng lint",
     "e2e": "npx ng e2e",
     "postinstall": "ngcc",
-    "test-dev-waf": "npx cypress run -c baseUrl=https://dev.dockstore.net -s \"cypress/integration/manualTests/waf.ts\"",
-    "test-staging-waf": "npx cypress run -c baseUrl=https://staging.dockstore.org -s \"cypress/integration/manualTests/waf.ts\"",
-    "test-prod-waf": "npx cypress run -c baseUrl=https://dockstore.org -s \"cypress/integration/manualTests/waf.ts\"",
-    "test-dev-no-auth": "npx cypress run --record --reporter junit -c baseUrl=https://dev.dockstore.net -s \"cypress/integration/manualTests/monitor.ts,cypress/integration/manualTests/basic-enduser.ts,cypress/integration/manualTests/search.ts\"",
-    "test-staging-no-auth": "npx cypress run -c baseUrl=https://staging.dockstore.org -s \"cypress/integration/manualTests/monitor.ts,cypress/integration/manualTests/basic-enduser.ts,cypress/integration/manualTests/search.ts\"",
-    "test-prod-no-auth": "npx cypress run -c baseUrl=https://dockstore.org -s \"cypress/integration/manualTests/monitor.ts,cypress/integration/manualTests/basic-enduser.ts,cypress/integration/manualTests/search.ts,cypress/integration/manualTests/basic-enduser-prod-only.ts\"",
-    "test-dev-auth": "npx cypress run --record --reporter junit -c baseUrl=https://dev.dockstore.net -s \"cypress/integration/manualTests/auth-tests.ts\"",
-    "test-staging-auth": "npx cypress run -c baseUrl=https://staging.dockstore.org -s \"cypress/integration/manualTests/auth-tests.ts\"",
-    "test-prod-auth": "npx cypress run -c baseUrl=https://dockstore.org -s \"cypress/integration/manualTests/auth-tests.ts\""
+
+    "test-shared-no-auth": "npx cypress run --record --reporter junit -s \"cypress/integration/smokeTests/sharedTests/monitor.ts,cypress/integration/smokeTests/sharedTests/basic-enduser.ts,cypress/integration/smokeTests/sharedTests/search.ts\"",
+    "test-shared-auth": "npx cypress run --record --reporter junit -s \"cypress/integration/smokeTests/sharedTests/auth-tests.ts\"",
+    "test-shared-waf": "npx cypress run -s \"cypress/integration/smokeTests/sharedTests/waf.ts\"",
+
+    "test-dev-waf": "npx cypress run -c baseUrl=https://dev.dockstore.net -s \"cypress/integration/smokeTests/devOnly/waf.ts\" && npm run test-shared-waf -- -c baseUrl=https://dev.dockstore.net",
+    "test-staging-waf": "npx cypress run -c baseUrl=https://staging.dockstore.org -s \"cypress/integration/smokeTests/stagingOnly/waf.ts\" && npm run test-shared-waf -- -c baseUrl=https://staging.dockstore.org",
+    "test-prod-waf": "npx cypress run -c baseUrl=https://dockstore.org -s \"cypress/integration/smokeTests/prodOnly/waf.ts\" && npm run test-shared-waf -- -c baseUrl=https://dockstore.org",
+
+    "test-dev-no-auth": "npx cypress run --record --reporter junit -c baseUrl=https://dev.dockstore.net -s \"cypress/integration/smokeTests/devOnly/monitor.ts,cypress/integration/smokeTests/devOnly/basic-enduser.ts,cypress/integration/smokeTests/devOnly/search.ts\" && npm run test-shared-no-auth -- -c baseUrl=https://dev.dockstore.net",
+    "test-staging-no-auth": "npx cypress run -c baseUrl=https://staging.dockstore.org -s \"cypress/integration/smokeTests/stagingOnly/monitor.ts,cypress/integration/smokeTests/stagingOnly/basic-enduser.ts,cypress/integration/smokeTests/stagingOnly/search.ts\" && npm run test-shared-no-auth -- -c baseUrl=https://staging.dockstore.org",
+    "test-prod-no-auth": "npx cypress run -c baseUrl=https://dockstore.org -s \"cypress/integration/smokeTests/prodOnly/monitor.ts,cypress/integration/smokeTests/prodOnly/basic-enduser.ts,cypress/integration/smokeTests/prodOnly/search.ts,cypress/integration/smokeTests/prodOnly/basic-enduser-prod-only.ts\" && npm run test-shared-no-auth -- -c baseUrl=https://dockstore.org",
+
+    "test-dev-auth": "npx cypress run --record --reporter junit -c baseUrl=https://dev.dockstore.net -s \"cypress/integration/smokeTests/devOnly/auth-tests.ts\" && npm run test-shared-auth -- -c baseUrl=https://dev.dockstore.net",
+    "test-staging-auth": "npx cypress run -c baseUrl=https://staging.dockstore.org -s \"cypress/integration/smokeTests/stagingOnly/auth-tests.ts\" && npm run test-shared-auth -- -c baseUrl=https://staging.dockstore.org",
+    "test-prod-auth": "npx cypress run -c baseUrl=https://dockstore.org -s \"cypress/integration/smokeTests/prodOnly/auth-tests.ts\" && npm run test-shared-auth -- -c baseUrl=https://dockstore.org"
   },
   "private": true,
   "dependencies": {
